analysis, rating priority, popularity formula
Plot analysis (more progress)
- Relevant debug command: /plot debugexec analyze
- I've been working on some more interesting plot analysis which will in
future give a better indication of quality content rather than blocks
changed
- The main problem with blocks changes is that if a plot is entirely set
to a blob of stone it will pass that requirement, despite the fact that
it's a blob of stone
- With the plot analysis it will calculate the complexity of the build
based on 13 different metrics
Rating priority:
- Relevant command to teleport to the next plot: /plot rate next
- PlotSquared will try to prioritize higher quality plots to receive
more ratings
- Low quality plots will still be rated, just not as much as higher
quality builds
Plot popularity:
- Relevent command to list popular plots: /plot list top
- Will sort the plots based on average rating and number of ratings

+++ PlotSquared/src/main/java/com/intellectualcrafters/plot/util/Permissions.java
@@ -38,10 +38,10 @@ public class Permissions {
 
     public static int hasPermissionRange(final PlotPlayer player, final String stub, final int range) {
         if ((player == null) || player.isOp() || player.hasPermission(ADMIN)) {
-            return Byte.MAX_VALUE;
+            return Integer.MAX_VALUE;
         }
         if (player.hasPermission(stub + ".*")) {
-            return Byte.MAX_VALUE;
+            return Integer.MAX_VALUE;
         }
         for (int i = range; i > 0; i--) {
             if (player.hasPermission(stub + "." + i)) {
